Do not set initial value
initial_value property of gpiozero LED class was set to "off".
This is an issue, during OctoPrint restart, for relays controlling
the printer power or even the OctoPrint itself (such as in my case)
because the plugin turns the pin off momentarily, cutting the power.
Using "None", the LED will be left in whatever state the pin is found
in when configured for output (quoted from gpiozero LED documentation)
and the initial value desired by the user will be applied immediately
after inizialization by the "on Startup" event, for a seamless UX.